fix animations duration

+++ Omnom.Linker/app/src/main/java/com/omnom/android/linker/utils/AnimationBuilder.java
@@ -25,7 +25,7 @@ public class AnimationBuilder {
 	private AnimationBuilder() {
 	}
 
-	public void setDuration(int duration) {
+	public void setDuration(long duration) {
 		animator.setDuration(duration);
 	}
